+++ client/UIFramework/SDL_Extensions.h
@@ -40,7 +40,7 @@ bool isItIn(const SDL_Rect * rect, int x, int y);
 
 namespace Colors
 {
-	SDL_Color createColor(int r, int g, int b, int a = 0);
+	SDL_Color createColor(Uint8 r, Uint8 g, Uint8 b, Uint8 a = 0);
 
 	const SDL_Color	Jasmine = createColor(229, 215, 123, 0); // http://en.wikipedia.org/wiki/Jasmine_%28color%29
 	const SDL_Color Cornsilk = createColor(255, 243, 222, 0); // http://en.wikipedia.org/wiki/Shades_of_white
